requirements.txt
psycopg2
streamlit
pandas
supabase
numpy 
scikit-learn
joblib
python-dotenv
langchain-google-genai
langchain
langchain-community